+++ app/src/main/java/edu/up/cs/androidcatan/gameframework/infoMsg/IllegalMoveInfo.java
@@ -1,0 +1,15 @@
+package edu.up.cs.androidcatan.gameframework.infoMsg;
+
+/**
+ * The a message from the game to a player that the move just attempted
+ * was illegal.
+ *
+ * @author Steven R. Vegdahl
+ * @version July 2013
+ */
+public class IllegalMoveInfo extends GameInfo {
+
+    // to satisfy Serializable interface
+    private static final long serialVersionUID = 7165334825841353190L;
+
+}
