+++ app/src/main/java/edu/up/cs/androidcatan/gameframework/util/Tickable.java
@@ -1,0 +1,19 @@
+package edu.up.cs.androidcatan.gameframework.util;
+
+/**
+ * An inteface that defines an object that can respond to a GameTimer "tick".
+ *
+ * @author Steven R. Vegdahl
+ * @version July 2013
+ *
+ */
+public interface Tickable {
+    /**
+     * callback method that is invoked when the timer "ticks"
+     *
+     * @param timer
+     * 		the timer that is associated with the "tick"
+     */
+    public abstract void tick(GameTimer timer);
+}
+
